Added utoa_base fct and changed all unsigned int, unsigned long, long, and unsigned char, by (u)int{8;32;64}_t

+++ ft_int_size.c
@@ -6,15 +6,15 @@
 /*   By: jodufour <jodufour@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/21 06:22:40 by jdufour           #+#    #+#             */
-/*   Updated: 2021/03/23 18:03:12 by jodufour         ###   ########.fr       */
+/*   Updated: 2021/05/10 01:29:44 by jodufour         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <sys/types.h>
+#include <stdint.h>
 
-size_t	ft_int_size(int nbr)
+uint32_t	ft_int_size(int nbr)
 {
-	size_t	size;
+	uint32_t	size;
 
 	size = 1 + (nbr < 0);
 	nbr /= 10;
